requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.18.0
pillow>=8.0.0
matplotlib>=3.3.0
opencv-python>=4.5.0
numpy>=1.19.0
pytest>=6.0.0